Structure-Centric-FG-SBIR/requirements.txt

torch>=1.13
torchvision
numpy
scipy
pyyaml
tqdm
